+++ src/controllers/auth/register.js
