+++ src/public/demo.html
@@ -40,8 +40,7 @@
     <div class="row">
       <div class="full column">
         <p>
-          If there is no one else to play with you, then launch another tab in private mode. 
-          Or use another device. To move, use the scrollbars or scroll with touchpad. Works best with Chrome or Firefox on laptop/computer.<br /><br />
+        If there is no one else to play with you, then launch another tab in private mode. Or use another device. To move, use the scrollbars or scroll with touchpad. Works best with Chrome or Firefox on laptop/computer. If things are glitchy, that's because I'm using HTML DOM to make the game. I'm not experienced in web-based gaming, so things will be a bit rough. To see the API in action, open up the web console and select the network tab.<br /><br />
           <form onsubmit="onSubmit(this); return false">
             <label for="username">Pick a username (alphanumberic): </label>
             <input type="text" id="username" pattern="[a-zA-Z0-9][a-zA-Z0-9]*" maxlength=20 /></br>
